ConfirmOrderController Initial Commit

+++ src/main/java/com/hcl/stocktrading/dto/ConfirmOrderDto.java
@@ -1,11 +1,13 @@
 package com.hcl.stocktrading.dto;
 
+import lombok.Data;
+
 @Data
 public class ConfirmOrderDto {
 
 	Long userId;
 	Long stockId;
-	Double purchaseVolume;
+	Long purchaseVolume;
 	
 	
 }
